[minor] move initAgentID and createIDFile to agentcontext
This allows these functions used within the agent to be used by the
loader as well. This is intended to establish a correlation between the
agent and a loader instance using the unique agent ID.
